Update code to allow an extra letter
Now, letters that aren't already being used in the cards get added to their own separate array. When a card is invalid, the main function adds one of those extra letters and tests if any of the permutations are a valid word, and if so, the letter is removed from the array so that it doesn't get used more than once. Since it's only used one time or less, it doesn't affect the chain rule, but it increases the likelihood that a card will be valid.