Exclude Q, X, and Z from cards
Including Q in any randomizations is a waste of time because no two words will have only Q in common, they will also both have U. So now Q is always at the end of the array and won't be included except as an extra letter. Z and X are also excluded for good measure, as those also lower the likelihood of a valid set.